The address of an instruction should not be relative
If an instruction is part of bank 1 it should be

0x4000 <= address <= 0x7fff

Before this they were

0x0000 <= address <= 0x4000

Which is only true for bank 0. It's necessary to fix this because we
can't offset for this in Memory.getByte. There's no way to know if an
instruction is trying to read the next byte during disassembly or if
something like LD ($0x2000), A is happening.